HPCC-13830 Split cassandraembed.cpp source file
